[TASK] Documentation of compressJs feature
The documentation (NEWS.txt and TypoScript completion) does not
correctly explain the feature set of config.compressJs.

Instead of the advertised minification (stripping white-spaces), it only
applies GZIP compression (which then requires compressionLevel to be
enabled in the Install Tool). This is a difference to
config.compressCss, which does such a minification.

Additionally, .htaccess lacks hints about the gzip rules, which now also
apply to the Frontend.

Change-Id: I30929ee70d0ab5fd6db74717889af6104c12e23d
Resolves: #31251
Releases: 4.6