[TASK] Clean up about modules extension
tx_aboutmodules_Functions was used in the good old days to render the main
module menus. Now, the only code that used this class was the about module
extension. The patch refactors the about module and moves the needed code of
the tx_aboutmodules_Functions to its own class. This makes the extension
way easier to maintain.
tx_aboutmodules_Functions is declared obsolete and unmaintained.

Change-Id: I15402c995a00f16006e20ecc40bd841d6b42c1cb
Resolves: #31296
Releases: 4.7
Reviewed-on: http://review.typo3.org/6338
Reviewed-by: Oliver Hader
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn