[FEATURE] Speed-up Backend with CardLayout
This patch adds a cardlayout to the Backend.
BE Modules may register themselves to the page renderer and add cards
directly to the layout.

To achieve that, a new way of adding JavaScript to the BE is available.
$GLOBALS['TBE_MODULES']['_JSINIT']['key'] = 'JSCODE';

This way changing the module is quite fast.

To keep old behaviour the iframe is loaded with the module root
url if the module item of an already opened module is clicked.

The Extension Manager or the About module can be simply reworked to support
that feature.

Other new features like the dashboard depend on this change too.

API hints:

JavaScript
 - in old times we used:
   TYPO3.Backend.ContentContainer.setUrl(uri);

   Now please use
   TYPO3.ModuleMenu.App.openInContentFrame(uri, params);
   The "new" function was available before and knows the current
   module. This function uses the correct iframe - if it is allowed to
   load the module.

PHP
 - added new Parameter to t3lib_extMgm::addModule to disable the card
   autocreation on demand

Change-Id: I6516366dbd7a03ec6bff9ca6138663405933a98c
Related: #30645
Releases: 4.7