[BUGFIX] Tx_Fluid_ViewHelpers_Be_Buttons_IconViewHelper broken
It seems as the sprite keys in
$GLOBALS['TBE_STYLES']['spriteIconApi']['iconsAvailable']
have changed with TYPO3 4.6 and this change was not reflected in
the experimental backend IconViewHelper, so the old key are not
longer valid and new ones cant be used because of the allowedIconArray.

I removed the "securityCheck" and the allowedIconArray, because the validity
of the icon name is checked again in t3lib_iconWorks::getSpriteIcon
against the iconAvailable array.

Change-Id: Ib7cebd9c6dbcd60d487480dea6bf263c9164c2fd