[BUGFIX] Sorting of localized records is wrong after copying
Solution: When copying a record in original language, if a localized
record of the destination record exists for language A, then the
localized record of the copied record, if it exists for language A,
should be inserted after the localized record, in language A, of the
destination record.

The same applies when moving a record in original language.

Change-Id: I78f1703db536bfabb64087f5e3e2064dc09b9eb2
Resolves: #30469
Releases: 4.5, 4.6, 4.7