[BUGFIX] DateTime properties are not mapped correctly
With the new property mapper *disabled* properties of type
DateTime are not correctly mapped to DateTime objects.
The reason is that the Object Container dies with an
exception when trying to reflect the (internal) constructor
arguments of DateTime objects.
This change fixes this by circumventing the Object Container
for DateTime objects.

Change-Id: I3db596a949a14941e89d67ecaa7a0606f5a80231
Resolves: #31345