[BUGFIX] DateTime properties are not mapped correctly
With the new property mapper *disabled* properties of type
DateTime are not correctly mapped to DateTime objects.
The reason is that the Object Container dies with an
exception when trying to reflect the (internal) constructor
arguments of DateTime objects.
This change fixes this by circumventing the Object Container
for DateTime objects.

Change-Id: I4166c895d983cdc2288d939a0d461dde31e5ee00
Resolves: #31345