[BUGFIX] Fatal error when exporting from root
An uncaught exception happens when trying to export records from a
singe table view in List mode on the page tree root. This is due to
the tree root being handled as a normal page record.

With this patch the tree root is handled as it should. Some cleanup
(CGL) is also applied.

Change-Id: I39fb872dd87d4f758695bbcb1607028506227fc6
Fixes: #31350
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/6791
Reviewed-by: Oliver Hader
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer