[FEATURE][API] Add PostProcessing Hook after inclusion of extTables
For modifiying the TCA of all Tables, you need to have an entry point
somewhere, shortly after all TCA definitions (which sould be within
ext_tables.php of each extension) are included.

To solve this problem a hook is introduced which allows to e.g. modify
all TCA tables in comparison to the classic way, where extensions are
only featured to modify TCA of extensions which had been loaded before.

Change-Id: Icd9653fccb477385f4c7107bdea637f6720c3d69
Resolves: #31615
Releases: 4.7
Reviewed-on: http://review.typo3.org/6564
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader
Reviewed-by: Dominik Mathern
Tested-by: Dominik Mathern
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter