[BUGFIX] Invalid query part on menu rendering
Each page can have a target defined, which can also be a
typeNum that gets added to the generated URL. Now if RealURL
is used, the "&type=" part will be just added to the final
URL which looks like "page.html&type=1". A question mark
is missing here to get a valid query part.

Change-Id: Idb5df6c604d12670107cbdd77c14e181cdb25dbe
Fixes: #31622
Releases: 4.7, 4.6, 4.5, 4.4