[BUGFIX] Invalid query part on menu rendering
Each page can have a target defined, which can also be a
typeNum that gets added to the generated URL. Now if RealURL
is used, the "&type=" part will be just added to the final
URL which looks like "page.html&type=1". A question mark
is missing here to get a valid query part.

Change-Id: I5404dde8963f2d5fea6a7e680cbe0f1d1f709d86
Fixes: #31622
Releases: 4.7, 4.6, 4.5, 4.4