[BUGFIX][!!!] Slider wizard misuses "max" property
TYPO3 4.6 introduced a slider wizard for select-type and
input-type fields. For input-type fields it is possible
to define a minimum and maximum value for the slider.
Unfortunately the wizard misuses the "max" TCA property
for the maximum value, which is normally used for the
"maxlength" attribute of the input field.

The slider wizard should instead rely on the "range" property,
which already exists and serves exactly the purpose of defining
a minimum and maximum value for the field.

Change-Id: I4bafc56d1cb0ce219c22e36f9998febb6c9b3675
Fixes: #31724
Releases: 4.7, 4.6
Reviewed-on: http://review.typo3.org/6628
Reviewed-by: Francois Suter
Tested-by: Francois Suter