[BUGFIX][!!!] Slider wizard misuses "max" property
TYPO3 4.6 introduced a slider wizard for select-type and
input-type fields. For input-type fields it is possible
to define a minimum and maximum value for the slider.
Unfortunately the wizard misuses the "max" TCA property
for the maximum value, which is normally used for the
"maxlength" attribute of the input field.

The slider wizard should instead rely on the "range" property,
which already exists and serves exactly the purpose of defining
a minimum and maximum value for the field.

Change-Id: I0cf34c78f5109ce244357471a122be0f24f1aeef
Fixes: #31724
Releases: 4.7, 4.6