[BUGFIX] Validation for multiple forms on one page
Since the class tx_form_System_Validate is a singleton, it was not
possible to use more than one form on a page. This patch changes
the arrays $rules, $messages and $errors, using the current form-
prefix as key, allowing for validation of several different forms
on one page.

Change-Id: I681c509c8884093dfe6ead527d2b44264f433fa6
Fixes: #31711
Releases: 4.7, 4.6
Reviewed-on: http://review.typo3.org/8920
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers