[BUGFIX] Validation for multiple forms on one page
Since the class tx_form_System_Validate is a singleton, it was not
possible to use more than one form on a page. This patch changes
the arrays $rules, $messages and $errors, using the current form-
prefix as key, allowing for validation of several different forms
on one page.

Change-Id: If4842a77fd7bfb3da62dbc3d79de01be88c708fd
Fixes: #31711
Releases: 4.7, 4.6