[FEATURE] More detailed felogin password reset error messages
This feature allows an error message to be displayed if the entered
account was not found.

The previous mechanism displayed a generic error message regardless
of a correct account name. The new mechanism allows a distinct error message if the account
was not found, which can be enabled by setting the "detailedForgotPasswordErrorMessage"
to 1 (default is 0). If enabled, instead of always displaying the
"ll_forgot_reset_message_emailSent" label, the "ll_forgot_reset_message_error" label will
be displayed if the entered account could not be found.

* Adds Typoscript option to enable/disable the new error message
  behavior (detailedForgotPasswordErrorMessage)
* Adds a new label "ll_forgot_reset_message_error" for the error message

Change-Id: I8d8ad52bf12938645bb9b144872ec64f92f875d0