[FEATURE] More detailed felogin password reset error messages
Allows a more explicit error message to be displayed, if the entered
account was not found.

The previous mechanism displayed a generic error message regardless of a
correct account name. The new mechanism allows a distinct error message
if the account was not found. This can be enabled by setting
"detailedForgotPasswordErrorMessage" to 1 (default is 0).

If enabled, instead of always displaying the message that an email was
sent, the user gets feedback that the account does not exist.

Change-Id: I8d8ad52bf12938645bb9b144872ec64f92f875d0
Resolves: #20911
Resolves: #23199
Releases: 4.7