[BUGFIX] Fix records already translated outside of IRRE
If you translate child records of IRRE-parents in the normal list-view,
then these child records won't be available in IRRE for translation
any more and causing these translated child records to be never
editable by IRRE again.
This bugfix adds already translated child records of an IRRE-parent
to the parent again, so the missing references will be adjusted.

Change-Id: I5253e992f94b1f9471872db25c0fe402b4551458
Fixes: #31812
Releases: 4.7, 4.6, 4.5, 4.4