[BUGFIX] Re-Include already translated records to IRRE parent, if child was translated outside of IRRE
If you translate child records of IRRE-parents in the normal list-view,
then these child records won't be available in IRRE for translation
any more. And so, these translated child-records will never be
editable by IRRE again.
This bugfix adds alread translated child-records of an IRRE-parent
to the parent again, so the missing references will be adjusted.

Change-Id: I5253e992f94b1f9471872db25c0fe402b4551458
Fixes: #31812
Releases: 4.7, 4.6, 4.5, 4.4