[FEATURE] Allow XCLASSes to be defined via ext_autoload.php
For XCLASSes to work currently, one needs to

a) set the according XCLASS statement in every PHP file that
   can/should be subclassed
b) use this XCLASS statement in an extensions' ext_localconf.php
   to include the subclass statement

This process is somehow unflexible, requires more code, and since
it is missing for some classes, it's also a hassle to maintain.

The attached change does the following:
1) Use the autoloader to search for classes prefixed by "ux_*"
2) Provide a backwards compatibility layer

TODO: Provide a possibility for BE admins to see which XCLASSes
are currently in use with this new method.

Change-Id: I4af4116108fecd9df0874c3e07e10b86f74d556e
Resolves: #31893
Releases: 6.0
Reviewed-on: http://review.typo3.org/6717
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert