[FEATURE] Allow XCLASSes to be defined via ext_autoload.php
For XCLASSes to work currently, one needs to

a) set the according XCLASS statement in every PHP file that
   can/should be subclassed
b) use this XCLASS statement in an extensions' ext_localconf.php
   to include the subclass statement

This process is somehow unflexible, produces code, and as it is
incomplete for some classes, a hassle to maintain.

The attached change does the following:
1) It makes use of the autoloader to search for classes of the
   XCLASS prefix "ux_*" as well.
2) It provides a way to also include the XCLASS paths via the
   "old way", but sends a deprecation warning for that.

There is one drawback however right now: There is no how an admin
can see which XCLASSes are currently in use with "the new way" as
the admin cannot browse the ext_autoload.php files of all
extensions through the backend in a nice way.

Also: Beware, you might need to clear your autoload cache when
trying out this feature, which you can do with a
simple "rm -rf Cache/Code/cache_phpcode/*"

Everything works as before now, even the limitation that only
one XCLASS can be used of course still applies. The XCLASS to
use is of course the one of the last extension installed,
same principle as before, just not in the ext_localconf.php
anymore, but in ext_autoload.php.

The advantages of "the new XCLASS way":

    "faster?" - Less initialization in ext_localconf.php,
    and only when it is needed
    "more flexible" - XCLASS every class in use now
    "more transparent" - no XCLASS statement on the bottom
    of each core class anymore

Change-Id: I4af4116108fecd9df0874c3e07e10b86f74d556e
Resolves: #31893