[BUGFIX] t3lib_BEfunc::getViewDomain() returns wrong or no domain
If the TYPO3 BE host name is not configured as a sys_domain record
t3lib_BEfunc::getViewDomain() will always return TYPO3 BE host name
and not the first domain of the root line of the page which is
being previewed.

Furthermore, if the TYPO3 BE host name is configured as a sys_domain
but the current root line does not contain a valid domain record,
t3lib_BEfunc::getViewDomain() will return only 'http://' ie. no domain.

Change-Id: I5290ac965234fad7e8980e84a13ec5877f5119fc
Fixes: #30892
Releases: 4.7, 4.6, 4.5