[TASK] Remove a message about non-existing TypoScript error
When using "< .xyz" TS constructs, TS object browser may show a
error like "1: Line 5902: Object copied in this line
"ACT < .NO" would leave either the value or properties untouched
in TypoScript Version 1. Please check that this is not a problem
for you.". The message is unclear and scary. It is shown as error
but there is, in fact, no error.

Change-Id: I51e811829cfe2d0f264861c4bb4adcdabf9d0546
Resolves: #32012
Releases: 4.7