[BUGFIX] Make generated thumbnails browser-cachable
Thumbnail generation code appends current timestamp in each
invocation of the thumbnail generation script. This is useless
because it prevents caching. Removing this timestamp will not
cause any issues with old vs new versions because the URL
includes an md5 checksum of the file size and file timestamp.
So when the file is updated, the md5 sum will change. The
patch also changes short md5 to full md5.

Additionally proper last-modified, expires and etag headers
are sent.

Change-Id: Icb32ca8fbde91dffa65b835ac9c0954f3f5da53f
Fixes: #21481
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/12820
Reviewed-by: Philipp Gampe
Reviewed-by: Marcus Schwemer
Reviewed-by: Wouter Wolters
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers