[BUGFIX] Make generated thumbnails browser-cachable
Thumbnail generation code appends current timestamp in each
invocation of the thumbnail generation script. This is useless
because it prevents caching. Removing this timestamp will not
cause any issues with old vs new versions because the URL
includes an md5 checksum of the file size and file timestamp.
So when the file is updated, the md5 sum will change. The
patch also changes short md5 to full md5.

Additionally proper last-modified, expires and etag headers
are sent.

Change-Id: I4e84f44e1e13ea5af2e9769e288f84eb1bf8f2ba
Fixes: #21481
Releases: 4.7, 4.6, 4.5