[BUGFIX] Generated tumbnails can't cached in the clients browser
Thumbnail generation code appends current time stamp to each
invocation of the thumbnail generation script. This is useless
because prevents caching. Removing this timestamp will not
cause any issues with old vs new versions because the URL
includes an md5 checksum of the file size and file timestamp.
So when the file is updated, the md5 sum will change. The
patch also changes short md5 to a full md5.

Change-Id: I852e6d8c4a2626c167fa4ea380874b46f17112ca
Resolves: #21481
Releases: 4.7