[FEATURE] Add error message for missing TCA
Adds an error message in case the TCA of a table is missing
column definitions. The error message adds a hint that helps
the developer to find the affected table.
In addition it changes the table header to show the table name
as it is in the database if there is no title defined in TCA.

Change-Id: I568c63d40de11385da0bf848904b860c4269fac4
Resolves: #17961
Releases: 4.7