[FEATURE] Add error message for missing TCA
Adds an error message in case the TCA of a table is missing
column definitions. The error message adds a hint that helps
the developer to find the affected table.
In addition it changes the table header to show the table name
as it is in the database if there is no title defined in TCA.

Change-Id: I2b9039fbde852e7967c4029c287e2a714ec3ffcf
Resolves: #17961
Releases: 4.7