[FEATURE] Add two hooks to tx_sysaction_task
Add central hook object registration in the sys_action task which gets
used for two hook methods. One that modifies the transfered data before
the process of the task self acts on them. The other hook that enables
custom error checking. Like special checks on the submitted password
for complexity or similier.

Change-Id: I2a2a2ffdfb896d798ce0b96c3d48334956cf49b7
Fixes: #32055
Releases: 4.7