[BUGFIX] PHP warning if open_basedir is enabled
When using flexforms, a warning may appear in the Backend when
saving the form in BE (typically in a plugin). This happens
because TYPO3 checks if flexform DS is static (in the file)
using is_file. When the DS is not static, the parameter
to is_file is not a file name but an XML content. PHP
thinks this "file name" falls outside of the open_basedir
and complains with a warning. Possibly it can be solved
in other ways, but this fix is the simplest one.

Change-Id: I2c95522418e74e68c96c223f19b022d300467f7d
Resolves: #32109
Releases: 4.7, 4.6, 4.5