[FEATURE] objective wrapper for t3lib_loadDBGroup
Currently you need to resolve realtions between TCA-tables
yourself or use t3lib_loadDBGroup for that.

The interface of t3lib_loadDBGroup is not object-orientated
and hard to use. Furthermore you need internal knowledge and
need to prepare data and configuration from "outside" before
using it.

To ease up the usage, encourage developers using core functionality
and doing this in a more object orientated way, the user feels
more comfortable with, add wrapper class dealing with these issues.

In the first step this acts only as a wrapper for t3lib_loadDBGroup
wherever needed but in longterm  t3lib_loadDBGroup should be
completely refactored into this relation service.

Change-Id: I8d0cd3fa302683e952f600041dad66e8a4a237bc
Resolves: #32148
Releases: 4.7