[TASK] Remove old code related to the extension tstemplate_styler
The extension tstemplate_styler does not come with TYPO3, it is
old amd most likely unused by anybody with a modern TYPO3 version.
However fields for this extension exist in TYPO3 tables as
TEXT/BLOB. Also code exists to manage the files for the styler
but this code is either under condition, which is always false,
or never called because corresponding files do not exist any more.
Since this functionality is unused, remove it without prior
deprecation in TYPO3 4.7.

Change-Id: Ic864276be13b069838f8950e43b864f1c12b3f75
Resolves: #26090
Releases: 4.7