[BUGFIX] Shortcut redirect ignores linkVars syntax
Currently the code that creates the redirect url for a shortcut
page only respects the linkVars option as list of parameters without
additional validation informations like it's documented in the
TSRef. Unfortunatly such paramaters are completely dropped and
this leads to major problems on multilanguage sites with the new
shortcut redirect feature.

Change-Id: I3966eedf284b06b7765f38ca07ea6cd7155babb3
Fixes: #32253
Releases: 4.7, 4.6
Reviewed-on: http://review.typo3.org/7015
Reviewed-by: Björn Pedersen
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Neufeind
Reviewed-by: Wouter Wolters
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch