[BUGFIX] Page within a mountpoint using Show Content from Page fatals
When pages:content_from_pid is used, the TSFE for the current object
will be cloned to a temporary object for the page to get the content
from.  In this temporary object, the id for the page will be changed to
the contents of content_from_pid.  However, the MP variable will not be
changed.

This results in a TSFE object with the wrong MP variable connected to
the id.  Within this temporary object the rootline will be calculated,
which will be empty, because of the wrong MP variable.  This triggers
"The requested page didn't have a proper connection to the tree-root"

This bugfix fixes this by emptying the MP variable.

Change-Id: I56fc25924a1766d2befaffee8abbed5466a04942
Fixes: #32292
Releases: 4.7, 4.6, 4.5
Reviewed-on: https://review.typo3.org/12837
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny