[BUGFIX] Page within mountpoint can result in 503 header
When pages:content_from_pid is used, the TSFE for the current
object will be cloned to a temporary object for the page to get the
content from. In this temporary object, the id for the page will be
changed to the contents of content_from_pid. However, the MP
variable will not be changed.

This results in a TSFE object with the wrong MP variable connected
to the id. Within this temporary object the rootline will be
calculated, which will be empty, because of the wrong MP variable.
This triggers the 503 header.

This bugfix fixes this by emptying the MP variable.

Change-Id: I785b41c0724b32f8ca740462da3101f4133a267b
Fixes: #32292
Releases: 4.7, 4.6, 4.5