[BUGFIX] Page within a mountpoint using Show Content from Page results in 503 header
When pages:content_from_pid is used, the TSFE for the current object will be cloned to a temporary object for the page to get the content from. In this temporary object, the id for the page will be changed to the contents of content_from_pid. However, the MP variable will not be changed.

This results in a TSFE object with the wrong MP variable connected to the id. Within this temporary object the rootline will be calculated, which will be empty, because of the wrong MP variable. This triggers the 503 header.

This bugfix fixes this by emptying the MP variable.

Change-Id: I380b33e6daf3d36a1799409f484cc2e75290681b
Fixes: #32292
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/7071
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Stefan Neufeind
Tested-by: Steffen Ritter