[BUGFIX] Page renderer provides files to compressor incorrectly
If config.absRefPrefix is set, page renderer provides incorrect
file references to compressor. It prepends absRefPrefix to JS and
CSS file names. This causes two problems: (1) compressor cannot
fetch those files because it always prepends site path to the
file name and (2) compressor fetches local files using HTTP
(performance!). The result is that all default styles from
extensions are missing when compressor is enabled but a
reference to an empty CSS file is added to the page.

Change-Id: I6be4f4066868533a230e8ac990d26874d5da55c4
Resolves: #32397
Releases: 4.7, 4.6