[BUGFIX] Invalid "unequal"-statement and string-quoting in SQL
There are multiple "!=" in SQL statements. They should be "<>".
MySQL understands "!=" for the compatibility with early versions
but it is not a standard SQL and should not be used.

Strings should be enclosed in single-quotes. Some systems use
double-quotes to mark identifiers in contrast to literals.

(Added some small fullQuote-calls along the way for
completeness or dropped quotes where uid is clearly just
an intval.)

Change-Id: If183e3451dac4cebb179b4c1f8c2e0ef6d5a6efd
Resolves: #32161
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/8900
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers