[FEATURE] Check for write-permissions before installation
If write-permissions for several directories are not properly set,
the installation process might fails as well. Especially if the
FIRST_INSTALL file is used, ENABLE_INSTALL_TOOL cannot be created
automatically in that case. Thus, a warning is issued if permissions
to the required directories are wrong and the ENABLE_INSTALL_TOOL
file was not found.

Change-Id: Ie48ce67f07f350a05c24d6d11bab962ecb11de1c
Resolves: #11771