[TASK] Add ExtJS3 Compatibility Layer Extension
Until all Backend-Modules are migrated to ExtJS4 we need to
make them available via ExtJS3. Thanks to Iframes this is possible.

This sysext ships all ExtJS 3 Code which had been in place before
the migration and adds an LoaderClass, which hooks into PageRenderer
to exchange the Library if requested.

Change-Id: Iacbbaff2217c7b04232ac31773f613d3913abb36
Resolves: #32720
Releases: 4.7