[BUGFIX] Pagerenderer is useless in USER_INT plugin
It looks like the pageRenderer actually renders all the page
content before the USER_INT plugins are substituted in the
cached output. When this substitution occurs, it seems
to only support the old TSFE->additionalHeaderData and
not the pageRenderer.

If you try per example to use tt_content.media.20 in a
USER_INT plugin, the JS library and the inline JS, is not
included, the patch solve the problem.

The patch is a port of the old patch available on forge,
tested on 4.5 and 4.6 with minor modification and refactoring.

Resolves: #22273
Change-Id: I97609684ccacdab1bd0853b9ccd6608716706f87