[BUGFIX] Make PageRenderer work with USER_INT plugins
Currently the PageRenderer renders all the page content before
USER_INT plugins are substituted in the cached output.
This leads to the situation, that adding header or footer data using
the PageRenderer does not work for USER_INT plugins.

If you for example try to use tt_content.media.20 in a
USER_INT plugin, the JS library and the inline JS, is not
included.

This change solves the problem, by adding the header and footer data
of the PageRenderer during USER_INT processing.

Resolves: #22273
Releases: 6.0
Change-Id: I97609684ccacdab1bd0853b9ccd6608716706f87
Reviewed-on: http://review.typo3.org/7465
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog