[BUGFIX] JavaScript error on new Flexform sections
The fix for #28952 causes a JavaScript error on newly created flexform
sections. As decodeURIComponent() already decodes the given URI it seems
to be unnecessary to apply unescape() to the result, so it could be
removed.

Applying this fix makes JS links on Flexform sections work again, utf8
labels fixed with #28952 are still working after removing unescape().

Change-Id: I18b8d77dbbdabe1e9767c12a41a8903df52fa7b1
Fixes: #32238
Releases: 4.7, 4.6, 4.5