[FEATURE] Add hooks to extend browse_links.php
Add hooks to be able to extend the browse links popup in RTE and for
TCEForms link fields.

This is useful to add new fields to the link wizard to add additional
params, onclick atrtributes, rel attributes or other tracking parameters
to the rendered link.

This requires hooks also in rtehtmlarea.

Change-Id: I917829cc3be423b4ae2e92a3264894152f912fcd
Fixes: #32809
Releases: 4.7