[BUGFIX] backColor transparent causes black background or flickering
When a GIFBUILDER object has set its backColor option to transparent,
the background may be black when adding a IMAGE object to it, and inner
areas of TEXT elements may be completly transparent, ignoring the
transparency by underlying elements.

This patch, removes the "imagealphablending" function calls, which
caused the latter issue and adds an additional check to the
tslib_stdGraphic::imagecopyresized method to prevent useage of the
imagecopyresized_fix workaround when working with transparency, as
it is responsible for the black background.

Backwards compatibility is not affected, since upgraded systems will
not have a backColor=transparent setting, thus have the property
saveAlphaLayer set to FALSE in any case.

Change-Id: I6fb22e3089689c7c61b2f04708504f6edb670ffa
Fixes: #32781
Releases: 4.7, 4.6
Reviewed-on: http://review.typo3.org/7622
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter