[CLEANUP] Core has a XML syntax dependency on templavoila
The TYPO3 core still uses "tx_templavoila" in a XML structure
for referencing the title for a structure like FlexForms with
sections etc. This means that code always looks / looked
like this:
<tx_templavoila>
<title>ABCDEF</title>
</tx_templavoila>

you can see this syntax in EXT:cms/flexform_media.xml.

The change in the Core is to support the new format
(without nesting in tx_templavoila) and deprecate but still
support the old format (see above).

Change-Id: I1af739858fb79299edcdbab9d4b44c063666aa9b
Resolves: #32946
Branches: 4.7
Reviewed-on: http://review.typo3.org/7663
Reviewed-by: Xavier Perseguers
Reviewed-by: Georg Ringer
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter