[FEATURE] Hidden palettes in TCEForms
This is useful for IRRE when used with foreign_selector, if you want to
hide the relation field in the child record.

By using the hidden palette feature you can just put it to such a hidden
palette, which makes sure IRRE will still work, but the relation field
is never displayed.

Change-Id: I78d04469ab1e902b98710dec633f8566814b4390
Resolves: #33044
Releases: 4.7