[BUGFIX] Scheduler BE module fails to load due to autoloader issues
When displaying a list of tasks, scheduler unserializes exceptions.
By default, unserialize() does not call autoloader to load classes
and it is possible that the exception class will not be loaded. In
such case scheduler BE module will fail with a PHP fatal error
when the scheduler will try to use a _PHP_Incomplete_Class.

The fix makes sure that autoloader is called during unserialize().
However this still may fail, so the scheduler now checks that the
exception is loaded properly before trying to use it.

Change-Id: I77ef3bf765019ab7ceb5bc45049d14b22db6f701
Resolves: #33116
Releases: 4.7, 4.6, 4.5