[BUGFIX] Regression in language handling
If you request a translated string in a
language where no translation is available
you used to get the default language
labels in the [target] array attribute of
the LOCAL_LANG array for that language.
After the fix of #32573 you only get the
[source] element filled. Which breaks
extensions (like extbase/fluid) that rely on
always having the [target] attribute filled.

Change-Id: I53edac0298e3415a7503259f25870e8dd92cddab
Resolves: #33227
Releases: 4.6, 4.7