[BUGFIX] Broken datepicker on unsaved flexform sections
The onClick event for opening the datepicker is attached to the
datepicker icons only once on page load. There currently is no update
method which adds the onClick event to new sections, so the datepicker
icons are without function until the form is saved.

Related to this, IDs of input fields are not unique when adding more than
one new section before saving. They have to be randomized when inserting a
new section in order to make the datepicker work properly.

This patch implements an update() method in tceforms.js which is called
every time a new section is added to a flexform. The IDs of tceforms
date/datetime fields get randomized by adding a timestamp when adding a
new section.

Fixes: #23843
Releases: 4.7, 4.6, 4.5

Change-Id: I9ac6e21b06bd1928c7acc94906113811f1260868
Reviewed-on: http://review.typo3.org/8674
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog