[BUGFIX] Draft usergroup access rights are now respected
Adding usergroup access rights to a page in a draft workspace had no
effect on the preview, the page was still shown in menus. The other way
around didn't work either. Removing access rights would not make the page
 available. Previewing wasn't possible at all.

This behavior is fixed. All draft usergroup changes are now respected
when previewing.

Change-Id: Ib9a74e98d6ed7457849465fc64685f89ed82fb61
Fixes: #33436
Releases: 6.0, 4.7, 4.6, 4.5