[BUGFIX] process_datamap doesn't handle existing versions by itself
The process_datamap function in TCEmain expects to receive the ids
of the existing offline counterparts of a record whenever operations
are performed in a workspace. This implies that this has taken into
account in all places where the process_datamap function is called.
In some cases this is not properly done, see bug #32780 and related.
The process_cmdmap function does handle this transparently,
so it is desirable to have the same behavior in process_datamap.

This fix implements the translation to versioned records for operations
on a single record. Operations with child records (IRRE) still require
the versioned id's, no bugs are known that would require this behavior
to be changed.

Change-Id: I1e41172b0cc1e04a4822147f3c48ca272b69467c
Fixes: #32780
Releases: 4.7