[BUGFIX] DB field names exceed max length with 'zzz_deleted'
MySQL allows a maximum of 64 characters for column and table names.
When prefixing them with 'zzz_deleted' this limit might be reached
by accident. (e.g. with
'zzz_deleted_tx_powermail_tellAFriendForm_receiverEmailContentField')

The patch avoids the overstepping 64 characters.

At the same time the prefixed and shortened field names are more
likely to have the same name, therefore the names are cut
from the beginning.

Change-Id: Ie2abe88be9ce38182a5d8db498fd27ff176c4cd0
Fixes: #33580
Releases: 4.5, 4.6, 4.7, 6.0
Reviewed-on: http://review.typo3.org/11327
Reviewed-by: Philipp Gampe
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn