[BUGFIX] DB field names exceed max length with 'zzz_deleted'
MySQL allows a maximum of 64 characters for column and table names.
When prefixing them with 'zzz_deleted' this limit might be reached
by accident. (e.g. with
'zzz_deleted_tx_powermail_tellAFriendForm_receiverEmailContentField')

The patch avoids the overstepping 64 characters.

At the same time the prefixed and shortened field names are more
likely to have the same name, therefore the names are cut
from the beginning.

Change-Id: I5fa03d17288090c9db6a47e097bbca0ce583a6e4
Fixes: #33580
Releases: 4.5, 4.6, 4.7, 6.0