[BUGFIX] Handling of parameterized headers in the SwiftMailerAdapter
SwiftMailer expects parameters to be set independent of the header
value itself. But currently the code doesn't actually split the raw
header value string but just passes it on as the value. That means that
SwiftMailer won't see the embedded parameter and things break.

For example, the charset parameter of the Content-Type header gets
duplicated, because SwiftMailer sets the parameter, because it thinks it
is missing.

Change-Id: I133208846ae10d805773d952c7f31bebc32d5520
Fixes: #33676
Releases: 4.7, 4.6, 4.5