[BUGFIX] Fix handling of Content-Transfer-Encoding in the SwiftMailerAdapter
The SwiftMailerAdapter sets the Content-Transfer-Encoding header in the
SwiftMailer message it creates, but fails to update the user encoder.

This means that the default quoted-printable encoder is used, which of
course breaks for all other encodings.

Therefore we need special handling for this header, in order to set the
encoder in addition to the plain header.

Change-Id: Id7ae6c7fe0a0623cf46e6006de3b77f801716cff
Fixes: #33633
Releases: 4.7, 4.6, 4.5