[BUGFIX] JavaScript error on new Flexform sections
The fix for #28952 causes a JavaScript error on newly created flexform
sections. As decodeURIComponent() already decodes the given URI it seems
to be unnecessary to apply unescape() to the result, so it could be
removed.

Since decodeURIComponent does not work on non utf-8 characters
this function is only used if the backend is rendered in utf-8.
In all other cases unescape is used.

Change-Id: I1d272b7d936062c55937ce54fb346350c9fcae93
Fixes: #32238
Releases: 4.7, 4.6, 4.5