[BUGFIX] Add the title and increase width of the multiple select boxes
The db, file and select boxes are too narrow. This makes it hard to
to impossible to read long entries.
Increase the width of those fields and add a title to each entry in
order to allow the user to see the full entry on hover.

Also add a parameter to the javascript function that takes care of the
selection of entries to support the new title attribute.

Change-Id: Ica18813c293e1d3c7a5e1447aad41b6ed5f35446
Fixes: #32732
Releases: 4.7,4.6,4.5
Reviewed-on: http://review.typo3.org/8978
Reviewed-by: Thomas Layh
Tested-by: Thomas Layh
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer