[BUGFIX] Added the title to the option fields and increased the width for the fields
The lists that contain the static includes (and available static templates)
in the template editing interface are too narrow. With some extensions it's
pretty much impossible to read the whole list item. There are no tooltips and
no (to me) obvious way to read the whole entry

Also added a parameter to the javascript function that takes care of the selection
of entries.

Change-Id: Id2161c2bda1b2d562ced4f5462c772bdfb412276
Fixes: #32732
Releases: 4.7,4.6,4.5