[BUGFIX] Add the title and increase width of the multiple select boxes
The db, file and select boxes are too narrow. This makes it hard to
to impossible to read long entries.
Increase the width of those fields and add a title to each entry in
order to allow the user to see the full entry on hover.

Also add a parameter to the javascript function that takes care of the
selection of entries to support the new title attribute.

Change-Id: Id2161c2bda1b2d562ced4f5462c772bdfb412276
Fixes: #32732
Releases: 4.7,4.6,4.5