[BUGFIX] show uid of reference in show_item.php
The uid of referenced records is not shown anymore and if there are
more items with same title, there is no easy way to identify the
record in the backend.

Solution is to add a column for the uid and page information in the
title attribute of the uid.

Change-Id: I998a87cac7c01df50b8bc884794b0418e967a191
Fixes: #33866
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/9005
Reviewed-by: Sebastian Fischer
Reviewed-by: Georg Ringer
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers