[BUGFIX] show uid of reference in show_item.php
The uid of referenced records is not shown anymore
and if there are more items with same title there
is no easy way to identify the record in the backend.

Solution is to just add the uid and pid in the title
attribute of the record title. This is applied also
for the table title since if the record title is
long it might have another title attribute holding
uncropped title.

Change-Id: I90c44cbff7840888c397d6667ae0a8efa0f5d0a7
Fixes: #33866
Releases: 4.7, 4.6, 4.5