[BUGFIX] show uid of reference in show_item.php
The uid of referenced records is not shown anymore and if there are
more items with same title, there is no easy way to identify the record
in the backend.

Solution is to add a column for the uid and page information in the
title attribute of the record title.

Change-Id: I90c44cbff7840888c397d6667ae0a8efa0f5d0a7
Fixes: #33866
Releases: 4.7, 4.6, 4.5