[BUGFIX] backColor transparent causes black background or flickering
When a GIFBUILDER object has set its backColor option to transparent,
the background may be black when adding a IMAGE object to it, and inner
areas of TEXT elements may be completly transparent, ignoring the
transparency by underlying elements.

This patch, removes the "imagealphablending" function calls, which
caused the latter issue and adds an additional check to the
tslib_stdGraphic::imagecopyresized method to prevent useage of the
imagecopyresized_fix workaround when working with transparency, as
it is responsible for the black background.

Backwards compatibility is not affected, since upgraded systems will
not have a backColor=transparent setting, thus have the property
saveAlphaLayer set to FALSE in any case.

Change-Id: I1871413a776ccd108e83c887105151b03bc1b667
Fixes: #32781
Releases: 6.0, 4.7, 4.6