[BUGFIX] Add the title and increase width of the multiple select boxes
The db, file and select boxes are too narrow. This makes it hard
to to impossible to read long entries. Increase the width of those
fields and add a title to each entry in order to allow the user to
see the full entry on hover.
Also add a parameter to the javascript function that takes care of the
selection of entries to support the new title attribute.

Change-Id: Ib03d45ac5884a6e5e91b9990043e1046ce69406e
Fixes: #32732
Releases: 4.7,4.6,4.5