[CLEANUP][CONF!!!] Remove forceCharset option and references
In TYPO3 4.5 the forceCharset and setDBinit options have been
deprecated. "forceCharset" was set to be removed in 4.7.

This ChangeSet removes the config option from the install tool
and all references within core to this variable.

If behaviour had been in question, the code has been adapted
to behave like the variable had been set to utf-8 and therefore
like the behaviour had been since 4.5.

Change-Id: Ic8fa9385b59eb127271c4f1a0902bbe8f3ee5499
Resolves: #34092
Releases: 4.7