[TASK] change charset fallbacks on iso-8859-1 to utf-8
Many places in core have characterset fallbacks if forceCharset
or corresponding settings are defined.

Starting with 4.5 utf-8 has been named default, but not changed
and set in all files, but just overridden with an hardcoded forceCharset.

As 4.7 removes this settings finally, all fallbacks in core to iso-8859-1
have to be rewritten to utf-8.

Change-Id: I9c228821e95167b67811c8475880707d5c77bdb7
Resolves: #34094
Releases: 4.7