[BUGFIX] Duplicate cache entry possible
The PdoBackend checks if the cache entry exists before it removes the entry when calling "set". The entry will not be found if the cache entry is expired. If the cache entry is expired it won't remove the entry but allow to add a new one with the same identifier, believing it does not exist.

This patch removes the check. The "set"-function will always try to remove the entry like the DbBackend does before creating a new one.

Change-Id: Ife4e35223394805fc6d9e0b820a5d42789133f7b
Fixes: #34129
Releases: 4.7 4.6 4.5