[FEATURE] GroupedForViewHelper accepts property path as groupBy argument
This change allows the GroupedForViewHelper to accept not only a property/array key as
argument but also a property path. This reduces memory consumption for more complex
object graphs. Also, It enables a developer to create a custom algorithm (getter) to provide
the value to compare. The change is backwards compatible.
(this is backported from #9244)

Change-Id: Idcc002e89f6ec121fcbd9e75cd002ff7ad2946f1
Related: #9244
Reviewed-on: http://review.typo3.org/9173
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn