[BUGFIX] Make tceform field placeholders IRRE compatible.
If the placeholder feature is used with IRRE relations,
e.g. sys_file_reference for FAL, the row-data passed into
getPlaceholderAttribute might be a multidimensional array and the
used placeholder key might be "__row|__foreign|name".

With the current implementations this is not respected and the
resulting placeholders will just output "Array" instead of the actual
field content.

Change-Id: Iaea2d53fecd29c4e0dc2453c2191e9a1aee137d4
Fixes: #33956
Releases: 4.7