[BUGFIX] Remove early call to enableFields
When previewing a draft shortcut, the enableFields function is called
before the TCA is loaded, resulting in a fatal error.

It seems that the enableFields call is not needed, as the state
of the original record is not important. If the original record would
have been deleted, then no draft version can exist.

Change-Id: I01161152450370fb1c5fd739b7ac3f73368bd385
Fixes: #32778
Releases: 4.7, 4.6, 4.5