[BUGFIX] Incorrect redirect with multiple user groups
If the "redirectMode" is set to "groupLogin" and there are several
groups with "felogin_redirectPid", the redirect should happen to
the first or last group (depending on the mode of "redirectFirst"
flag. The actual redirect result is undefined because felogin
fetches groups from the database without any sorting. Depending
on MySQL internal index state, results can come in any order and
this order can change over time. In addition felogin uses
$GLOBALS['TSFE']->fe_user->groupData['uid'], which contains sorted
uid values for groups. This prevents correct redirects for
multiple groups too.

Felogin should ensure that it adds redirects URLs in the proper order.

Change-Id: I51d6a836fcf07c66d45d4bd6e0fa4743570e52e5
Resolves: #34383
Releases: 4.5, 4.6, 4.7