[TASK] Remove clear cache by id from clearCacheCMD
This issue is related to #34352. There a feature was introduced
that allows to clear the cache by tag and by id from
TCEMAIN.clearCacheCmd.

As discussed (also check comments on https://review.typo3.org/#/c/9266/)
the clear by id should not be part of the public api as a cache
identifier is something internal that's not disposed to the
outside in case of cache_pages.

This patch removes the clear cache by id part
(while keeping the clear cache by tag).

Change-Id: I7eddf382bb9ae1611e7fd257f43d43259bdf7c8a
Resolves: #34474
Releases: 4.7, 4.8
Reviewed-on: http://review.typo3.org/9389
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter