[FEATURE] Add hook to t3lib_cache_frontend_VariableFrontend->set()
A hook is needed that is processed when storing something
into the cache.

This hook could be used to reuse the page tags in some external caches
like nc_staticfilecache or even varnish. In combination with the
hook introduced in #34475 external caches could easily be synced
to the TYPO3 page cache.

Change-Id: I0452b937c48ec5ab4e4147995da600bf9815f20a
Resolves: #34476
Releases: 4.8
Reviewed-on: http://review.typo3.org/9328
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter