[BUGFIX] Nested IRRE elements are not removed
Imagine a nested IRRE structure like Hotel->Offer->Price (taken
form EXT:irre_tutorial) in the live workspace. Now the whole
structure from the parent record including all child records is
deleted on a custom workspace. Publishing the delete placeholders
does not work for child records, only the parent gets removed and
the children stay...

Since the parent record gets processed first, the "delete" action
during publishing is forwared to all child records, but without
the accordant publishing context. Thus, for removals it is now to
be checked whether the record is marked to be removed and would
be processed in the same publishing request, but just later.

Change-Id: If5741f5354507938eb4ddbe2015112acc47bccce
Fixes: #34556
Releases: 4.5, 4.6, 4.7
Reviewed-on: http://review.typo3.org/9446
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader