[BUGFIX] Custom validators may be overriden by default validators
The softref parser always puts the type to 'db' for db records.
Due to the abstract validators fetchType method just returning this
type, the type gets overriden if any validator without own fetchType
method is evaluated later.
The solution is to store the type inside the record array, so that the
returned type is the last special type found.

Change-Id: I2a62dd65f3455333e7ab13f0c8cd0bb57459885f
Fixes: #34566
Releases: 4.5,4.6,4.7,6.0