[FEATURE] Create Postprocessor to save to Database
This is a basic concept, how this postprocessor could be implemented.
It is not yet finished. Currently it doesn't even work.

Since I had no time to do anything on it for one week and will not be
able to work on it for another week or more, I wanted to commit this
patch to get some feedback and be able to work on that when I come
back.

Is the overall concept correct? Should the basic design be altered?

Currently I am wondering, how rendering the CVS-view of the form entry
should be implemented.
Any help is appreciated.

Resolves: 33914
Versions: 4.6, 4.7, 4.8

TODO: render-method in view classes of elements to be changed.
+++++++++++++++++++++++++++++++++++++++++++++++++++++++

- Configure $TCA in ext_tables.php
	- add a new table or
	- use tt_content?
		http://typo3.org/documentation/document-library/core-documentation/doc_core_tca/current/
- Storage
	- store id of form record
	- store as cvs or
	- store as xml using t3lib_div::array2xml_cs and t3lib_div::xml2array?
	- how should file uploads be stored?

- Use $GLOBALS['TYPO3_DB'] in both frontend and backend to access DB (t3lib_DB)
	http://typo3.org/documentation/document-library/extension-manuals/dbal/current/
	$GLOBALS['TYPO3_DB']->exec_INSERTquery('tt_address', array('name' => $name, 'email' => $email));

- TS configuration
	- pid where the stuff should be stored / required
	- hidden / default to yes
	- folder for file uploads / default to uploads/form
	- header of record / default to 'Submitted data from form'
	- headerField of record
	- delimiter / default to '\t'

General reading:
http://typo3.org/documentation/document-library/core-documentation/doc_core_api/current

Change-Id: I17a1c9e71f78c67ca5d41d3ee1cfaf8b8c54a21e