[BUGFIX] Custom HTML tags no longer malformed in IE
Custom HTML tags would get malformed while editing in IE.
If a tag is not present in HTML4 and the document mode
of IE is lower than IE9 Standards, then the tag is not
present in the DOM and not parsed correctly.
This is fixed by specifying possible custom tags in TSconfig
and creating dummy instances of the tags before loading
the contents in the DOM, also known as the Shiv trick.

Change-Id: Ie828588b94e6ee01bd645b49a0787deca9e07a40
Fixes: #34786
Releases: 4.5, 4.6, 4.7, 6.0
Reviewed-on: http://review.typo3.org/10913
Reviewed-by: Bart Dubelaar
Reviewed-by: Stanislas Rolland
Tested-by: Stanislas Rolland