[BUGFIX] Custom HTML tags no longer malformed in IE
Custom HTML tags would get malformed while editing in IE.
If a tag is not present in HTML4 and the document mode
of IE is lower than IE9 Standards, then the tag is not
present in the DOM and not parsed correctly.
This is fixed by specifying possible custom tags in TSconfig
and creating dummy instances of the tags before loading
the contents in the DOM, also know as the Shiv trick.

Change-Id: Ie9ee09947a88dd73f92c0330f5d0c15609a7efb3
Fixes: #34786
Releases: 4.7, 4.6, 4.5