[SECURITY] Protect arguments of form __referrer with HMAC
The request arguments of the referring request are
a serialized string written to one of the hidden
fields in a Fluid form. This string has to be protected
by a HMAC to protect Extbase from possible unserialize
attacks.

Note: For now there is no object known within Extbase,
that could be used for an unserialize exploit!

Change-Id: I998b4e6b270aa3d389e772d1fcb7aa57283b4f29
Security-Bulletin: TYPO3-CORE-SA-2012-001
Related: #35310
Releases: 1.4, 4.7, 6.0