[SECURITY] Protect arguments of form __referrer with HMAC
The request arguments of the referring request are
a serialized string written to one of the hidden
fields in a Fluid form. This string has to be protected
by a HMAC to protect Extbase from possible unserialize
attacks.

Note: For now there is no object known within Extbase,
that could be used for an unserialize exploit!

Change-Id: Ic59b34bd9b58e43158ebe05116c8f577334a729e
Security-Bulletin: TYPO3-CORE-SA-2012-001
Related: #35310
Releases: 1.4, 4.7, 6.0