[BUGFIX] Shortcuts within mounted trees lose MP parameter
When mounting a tree which has shortcut pages in it, a link to such a
shortcut page has the MP parameter in it. But when redirecting to the
destination page of the shortcut, this MP parameter is lost, thus making
the page lose its connection to the tree root (or rather to the part of
the rootline above the mountpoint).

Change-Id: I2167a3a9c55c278aa7fca578c784b19c837fe5c9
Fixes: #17537
Branches: 4.5, 4.6, 4.7, 6.0