[BUGFIX] Shortcuts within mounted trees lose MP parameter
When mounting a tree which has shortcut pages in it, a link to such a
shortcut page has the MP parameter in it. But when redirecting to the
destination page of the shortcut, this MP parameter is lost, thus making
the page lose its connection to the tree root (or rather to the part of
the rootline above the mountpoint).

Change-Id: Ied673571c74d584848ca2b2ed53a4c81cae9948d
Fixes: #17537
Branches: 4.5, 4.6, 4.7, 6.0