[BUGFIX] Page cache expiry calculation fails
The calculation of the page cache expirationy timestamp based on
property config.cache.xx = table:yy fails in some specific
scenarios, name when there are records having either a start time
in the future but no end time, or an end time in the future
and no start time. Contrary to what is being done so far, calculation
must be performed separately for each time field to avoid
interferences.

Change-Id: I4e4e259083a10f7125760d14ede11ff6a68022fe
Fixes: #35684
Releases: 6.0, 4.7, 4.6
Reviewed-on: http://review.typo3.org/10156
Reviewed-by: Philipp Gampe
Reviewed-by: Wouter Wolters
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Xavier Perseguers
Reviewed-by: Francois Suter
Tested-by: Francois Suter